pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipeline"
version = "0.1"
requires-python = ">=3.12,<3.13"
dependencies = [
    "confluent_kafka>=2.9.0",
    "pyyaml>=6.0.2",
    "kafka-python>=2.1.2",
    "pyspark==3.5.5",
    "websocket-client>=1.8.0",
    "clickhouse-driver>=0.2.9"
]

[tool.setuptools]
packages = ["scripts"]
